+++ src/main/java/labbook_connect/labbook_connect/App.java
@@ -1,6 +1,7 @@
 package labbook_connect.labbook_connect;
 
 import java.net.InetSocketAddress;
+import java.util.List;
 
 import org.eclipse.jetty.server.Server;
 import org.eclipse.jetty.servlet.ServletContextHandler;
@@ -10,6 +11,7 @@ import org.glassfish.jersey.servlet.ServletContainer;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import plugin.Analyzer;
 
 public class App {
 	
@@ -17,9 +19,16 @@ public class App {
 	
 	public static final String VERSION  = "1.0.0";
 	public static final int NUM_VERSION = 1000;
+	
+	public static List<Analyzer> analyzers_classes;
+	public static List<Analyzer> analyzers_loaded;
 
 	public static void main(String[] args) {
 		logger.info("BEGIN App main");
+		
+		AnalyzerLoader analyzerLoader = new AnalyzerLoader();
+    	
+    	analyzerLoader.loadAnalyzers();
 		
 		String host = "0.0.0.0";
 		int port = 8080;
@@ -28,8 +37,6 @@ public class App {
 
 		// Set API REST
 		final ResourceConfig config = new ResourceConfig(MyResource.class);
-		//config.packages("labbook_connect.labbook_connect");
-		//config.register(MultiPartFeature.class);
 
 		// INIT server
 		Server server = new Server(serv_host);
@@ -51,5 +58,4 @@ public class App {
 			server.destroy();
 		}
 	}
-
 }