+++ CHANGELOG.md
@@ -3,7 +3,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
-## [1.0.0] - 2023-11-20
+## [1.0.0] - 2023-12-13
 ### Added
 - Run a Jetty server with REST API managed by Jersey Library
-- 
+- Logs in file labbook_connect.log
+- Makefile to build, run, stop or clean the container
+- HAPI HL7 library
+- analyzer plugin system with external JAR file
+- load instance of analyzer with TOML setting file